requirements.txt
aiohttp
python-dotenv
fastapi
uvicorn
pydantic
requests
pyaudio
tempfile
keyboard
wave